# backend/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: Comfort Curators V0 API Contract Floor
  version: 0.1.0
  description: >-
    Minimum public and operational REST surface for the Comfort Curators V0
    modular monolith. Domain requirements and authorization rules in the
    protected product files remain normative.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/properties/{property_id}/calendar-feeds:
    get:
      operationId: listCalendarFeeds
      tags: [Reservations]
      parameters: [{$ref: '#/components/parameters/PropertyId'}]
      responses:
        '200': {$ref: '#/components/responses/Collection'}
    post:
      operationId: createCalendarFeed
      tags: [Reservations]
      parameters: [{$ref: '#/components/parameters/PropertyId'}]
      requestBody: {$ref: '#/components/requestBodies/Command'}
      responses:
        '201': {$ref: '#/components/responses/Resource'}
        '422': {$ref: '#/components/responses/Error'}
  /v1/calendar-feeds/{feed_id}/polls:
    post:
      operationId: requestCalendarPoll
      tags: [Reservations]
      parameters: [{$ref: '#/components/parameters/FeedId'}]
      requestBody: {$ref: '#/components/requestBodies/Command'}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /v1/properties/{property_id}/reservations:
    get:
      operationId: listPropertyReservations
      tags: [Reservations]
      parameters: [{$ref: '#/components/parameters/PropertyId'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {$ref: '#/components/responses/Collection'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    PropertyId: {name: property_id, in: path, required: true, schema: {$ref: '#/components/schemas/Identifier'}}
...
    FeedId: {name: feed_id, in: path, required: true, schema: {$ref: '#/components/schemas/Identifier'}}
...
    Cursor: {name: cursor, in: query, required: false, schema: {type: string, maxLength: 512}}
    Limit: {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
...
  requestBodies:
    Command:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CommandEnvelope'}
...
  responses:
...
    Accepted:
      description: Durable command accepted
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Accepted'}
    Resource:
      description: Resource representation
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Resource'}
    Collection:
      description: Cursor-paginated resource collection
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Collection'}
    Error:
      description: Stable error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Identifier:
      type: string
      minLength: 16
      maxLength: 64
      pattern: '^[A-Za-z0-9_-]+$'
    CommandEnvelope:
      type: object
      additionalProperties: false
      required: [idempotency_key, data]
      properties:
        idempotency_key: {type: string, minLength: 8, maxLength: 128}
        reason: {type: string, maxLength: 1000}
        data: {type: object, additionalProperties: true}
...
    Accepted:
      type: object
      additionalProperties: false
      required: [request_id, status, resource_id]
      properties:
        request_id: {$ref: '#/components/schemas/Identifier'}
        status: {type: string, enum: [accepted, queued]}
        resource_id: {$ref: '#/components/schemas/Identifier'}
    Resource:
      type: object
      required: [id, version, data]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        version: {type: integer, minimum: 1}
        data: {type: object, additionalProperties: true}
    Collection:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Resource'}}
        next_cursor: {type: [string, 'null']}
    Error:
      type: object
      additionalProperties: false
      required: [request_id, code, message]
      properties:
        request_id: {$ref: '#/components/schemas/Identifier'}
        code: {type: string, pattern: '^[A-Z0-9_]+$'}
        message: {type: string, maxLength: 1000}
        details: {type: object, additionalProperties: true}